🎨 Refactor the Events Package
This commit completely overhauls the structure of the events package.
Various handler functions have been created to handle slack events.
These handler functions have been moved into their own files within the
events package for increased simplicity and readability as well.
