+++ internal/coffee/coffee.go
@@ -3,6 +3,7 @@ package coffee
 import (
 	"fmt"
 	"math/rand"
+	"strings"
 	"time"
 
 	"github.com/slack-go/slack"
@@ -15,56 +16,20 @@ type CoffeeSpot struct {
 	Location string
 }
 
-var mackayCoffeeSpots = []CoffeeSpot{
-	{Name: "9th Lane Grind", Locale: "Central", Location: "https://goo.gl/maps/EDPURuJkywUqwfL68"},
-	{Name: "Jamaica Blue (Sydney St)", Locale: "Central", Location: "https://goo.gl/maps/ypUuZY2Yhpry4JCV7"},
-	{Name: "Jamaica Blue (Caneland Central)", Locale: "Central", Location: "https://goo.gl/maps/ZG6YCQ8EEokgUw5XA"},
-	{Name: "Jamaica Blue (Mt Pleasant)", Locale: "northside", Location: "https://goo.gl/maps/eLPJLVEzvHC91GrS8"},
-	{Name: "Lava Coffee (Mt Pleasant)", Locale: "Northside", Location: "https://goo.gl/maps/h8cFw3E6d5tjsQ3Z9"},
-	{Name: "Gloria Jeans (Caneland Central)", Locale: "Central", Location: "https://goo.gl/maps/tuW8gy4t3WQixwB16"},
-	{Name: "Gloria Jeans (Mount Pleasant)", Locale: "Northside", Location: "https://goo.gl/maps/3pmZB9hGz72RAy859"},
-	{Name: "Gloria Jeans (North Mackay)", Locale: "Northside", Location: "https://goo.gl/maps/maM6vR5q5KdffBHg6"},
-	{Name: "Muffin Break (Caneland Central)", Locale: "Central", Location: "https://goo.gl/maps/sn8VWrZ8fHcWqkCL6"},
-	{Name: "The Deli Nook", Locale: "Central", Location: "https://goo.gl/maps/pjyK5NngSeNeSB568"},
-	{Name: "Chances Op & Coffee Shop", Locale: "Central", Location: "https://goo.gl/maps/EyxckqWfMPaPwoQ3A"},
-	{Name: "YAW", Locale: "Central", Location: "https://g.page/YAWFoods?share"},
-	{Name: "Charlie's Cafe", Locale: "Central", Location: "https://goo.gl/maps/BxkcbqAAwT6dnNrP8"},
-	{Name: "The Gallery Cafe & Co", Locale: "Central", Location: "https://goo.gl/maps/Ceo7oNQ6Fcnhk5BP8"},
-	{Name: "The Coffee Club (Wood St)", Locale: "Central", Location: "https://goo.gl/maps/V26N5ciH5Vr1eieY9"},
-	{Name: "The Coffee Club (Caneland Central)", Locale: "Central", Location: "https://goo.gl/maps/YfBNx9mPKGW2dFcZA"},
-	{Name: "The Grazing Goat", Locale: "Central", Location: "https://goo.gl/maps/KrZQ1z9Mt1ccaUuN8"},
-	{Name: "Dispensary", Locale: "Central", Location: "https://g.page/thedispensarymackay?share"},
-	{Name: "Oscar's Cafe & Bar", Locale: "Central", Location: "https://goo.gl/maps/dAVnSLAJ2jrPa6My6"},
-	{Name: "Foodspace", Locale: "Central", Location: "https://g.page/FoodspaceCafe?share"},
-	{Name: "Ador'a Cafe", Locale: "Central", Location: "https://goo.gl/maps/r88eUCzCJ59j5cfG8"},
-	{Name: "Stellarossa (Parkside)", Locale: "Southside", Location: "https://g.page/StellarossaParkside?share"},
-	{Name: "Stellarossa (Mt Pleasant)", Locale: "Northside", Location: "https://g.page/stellarossa-mt-pleasant?share"},
-	{Name: "Sage on Hamilton", Locale: "Central", Location: "https://goo.gl/maps/MHzKLgD7or8HyRrs6"},
-	{Name: "Botanic Gardens Cafe", Locale: "Southside", Location: "https://g.page/botanic-gardens-cafe-west-mackay?share"},
-	{Name: "Curb", Locale: "Nothside", Location: "https://goo.gl/maps/zQYd8Gd9wL2VsVWr7"},
-	{Name: "K&Co", Locale: "Northside", Location: "https://goo.gl/maps/qx19tgvxag37WcDS6"},
-	{Name: "Galleons Restaurant", Locale: "Southside", Location: "https://goo.gl/maps/a9iV22yhe3i6cub57"},
-	{Name: "Carlyle & River Coffee Co", Locale: "Southside", Location: "https://goo.gl/maps/yEw7Uo2hZ78wrE5eA"},
-	{Name: "Wake House", Locale: "Northside", Location: "https://goo.gl/maps/861zYQQEq93ZCdhV8"},
-	{Name: "Byrnes (Willetts Road)", Locale: "Northside", Location: "https://g.page/byrnes-willetts-road?share"},
-	{Name: "Byrnes (Andergrove)", Locale: "Northside", Location: "https://goo.gl/maps/jKwsyWeYTxfvjohx6"},
-	{Name: "Woodman's Axe Espresso Bar Mackay", Locale: "Central", Location: "https://goo.gl/maps/qnRLkje6xnJUZ5in8"},
-	{Name: "Primal Coffee Roasters", Locale: "Southside", Location: "https://goo.gl/maps/m3xtKdfv2ENjLo9y6"},
-}
-
 // Function which returns a random coffee shop from the list, optionally filtered by locale.
 func getRandomCoffeeShop(locale string) (CoffeeSpot, error) {
 	var cs []CoffeeSpot
-	// If a locale is specified, filter the list of coffee shops by that locale.
-	if locale != "" {
+
+	// If a locale not specified, randomly select from all coffee spots, otherwise, filter the list of coffee shops by that locale.
+	switch locale {
+	case "":
+		cs = mackayCoffeeSpots
+	default:
 		for _, s := range mackayCoffeeSpots {
 			if s.Locale == locale {
 				cs = append(cs, s)
-
 			}
 		}
-	} else {
-		cs = mackayCoffeeSpots
 	}
 
 	// Pick a random number between 0 and the length of the list.
@@ -86,11 +51,14 @@ func coffeeShopMessage(cs CoffeeSpot) string {
 		lm = fmt.Sprintf("Located on the *%s* side of town.\n", cs.Locale)
 	}
 
-	msg := fmt.Sprintf("Here's a random coffee shop for you to check out! :coffee:\n"+
-		"*%s*\n"+
-		"%s\n"+
-		"Check out the location on Google Maps: %s\n", cs.Name, lm, cs.Location)
-	return msg
+	var sb strings.Builder
+
+	fmt.Fprint(&sb, "Here's a random coffee shop for you to check out! :coffee:\n")
+	fmt.Fprintf(&sb, "*%s*\n", cs.Name)
+	fmt.Fprintf(&sb, "%s\n", lm)
+	fmt.Fprintf(&sb, "Check out the location on Google Maps: %s\n", cs.Location)
+
+	return sb.String()
 }
 
 // createCoffeeAttachment creates a Slack attachment to prompt user for locale preference.
@@ -99,33 +67,9 @@ func createCoffeeAttachment() slack.Attachment {
 		//Set the callback ID to "coffee" so we can identify the message later.
 		CallbackID: "coffee",
 		Text:       "Where About's in Mackay Would Best Suit?",
-		Actions: []slack.AttachmentAction{
-			{
-				Name:  "location",
-				Text:  "Northside",
-				Type:  "button",
-				Value: "Northside",
-			},
-			{
-				Name:  "location",
-				Text:  "Southside",
-				Type:  "button",
-				Value: "Southside",
-			},
-			{
-				Name:  "location",
-				Text:  "Central",
-				Type:  "button",
-				Value: "Central",
-			},
-			{
-				Name:  "location",
-				Text:  "I Don't Mind",
-				Type:  "button",
-				Value: "",
-			},
-		},
+		Actions:    actions,
 	}
+
 	return attachment
 }
